+++ src/api/machine/SoftwareDeliveryMachine.ts
@@ -21,10 +21,10 @@ import { Target } from "../../spi/deploy/Target";
 import { GoalSetter } from "../mapping/GoalSetter";
 import { PushRule } from "../mapping/support/PushRule";
 import { StaticPushMapping } from "../mapping/support/StaticPushMapping";
-import { EditorRegistration } from "./EditorRegistration";
+import { EditorRegistration } from "../registration/EditorRegistration";
+import { GeneratorRegistration } from "../registration/GeneratorRegistration";
 import { ExtensionPack } from "./ExtensionPack";
 import { FunctionalUnit } from "./FunctionalUnit";
-import { GeneratorRegistration } from "./GeneratorRegistration";
 import { GoalDrivenMachine } from "./GoalDrivenMachine";
 import { ListenerRegistration } from "./ListenerRegistration";
 import { SoftwareDeliveryMachineOptions } from "./SoftwareDeliveryMachineOptions";
